+++ conda_build/build.py
@@ -42,7 +42,6 @@ from . import __version__ as conda_build_version
 from . import environ, noarch_python, source, tarcheck, utils
 from .config import Config
 from .create_test import create_all_test_files
-from .deprecations import deprecated
 from .exceptions import (
     BuildScriptException,
     CondaBuildException,
@@ -76,7 +75,6 @@ from .utils import (
     CONDA_PACKAGE_EXTENSIONS,
     env_var,
     glob,
-    on_linux,
     on_mac,
     on_win,
     shutil_move_more_retrying,
@@ -1738,7 +1736,10 @@ def bundle_conda(
             if (
                 # WSL bash is always the same path, it is an alias to the default
                 # distribution as configured by the user
-                on_win and Path("C:\\Windows\\System32\\bash.exe").samefile(args[0])
+                on_win
+                # check if WSL is installed before calling Path.samefile/os.stat
+                and (wsl_bash := Path("C:\\Windows\\System32\\bash.exe")).exists()
+                and wsl_bash.samefile(args[0])
             ):
                 raise CondaBuildUserError(
                     "WSL bash.exe is not supported. Please use MSYS2 packages. Add "
@@ -2867,32 +2868,6 @@ def warn_on_use_of_SRC_DIR(metadata):
             )
 
 
-@deprecated(
-    "3.16.0",
-    "24.9.0",
-    addendum=(
-        "Test built packages instead, not recipes "
-        "(e.g., `conda build --test package` instead of `conda build --test recipe/`)."
-    ),
-    deprecation_type=FutureWarning,  # we need to warn users, not developers
-)
-def _construct_metadata_for_test_from_recipe(recipe_dir, config):
-    config.need_cleanup = False
-    config.recipe_dir = None
-    hash_input = {}
-    metadata = expand_outputs(
-        render_recipe(recipe_dir, config=config, reset_build_id=False)
-    )[0][1]
-
-    utils.rm_rf(metadata.config.test_dir)
-
-    if metadata.meta.get("test", {}).get("source_files"):
-        if not metadata.source_provided:
-            try_download(metadata, no_download_source=False)
-
-    return metadata, hash_input
-
-
 def _construct_metadata_for_test_from_package(package, config):
     recipe_dir, need_cleanup = utils.get_recipe_abspath(package)
     config.need_cleanup = need_cleanup
@@ -3032,18 +3007,7 @@ def _extract_test_files_from_package(metadata):
 
 
 def construct_metadata_for_test(recipedir_or_package, config):
-    if (
-        os.path.isdir(recipedir_or_package)
-        or os.path.basename(recipedir_or_package) == "meta.yaml"
-    ):
-        m, hash_input = _construct_metadata_for_test_from_recipe(
-            recipedir_or_package, config
-        )
-    else:
-        m, hash_input = _construct_metadata_for_test_from_package(
-            recipedir_or_package, config
-        )
-    return m, hash_input
+    return _construct_metadata_for_test_from_package(recipedir_or_package, config)
 
 
 def _set_env_variables_for_build(m, env):
@@ -3578,23 +3542,6 @@ def tests_failed(
             os.path.dirname(os.path.dirname(pkg)), verbose=config.debug, threads=1
         )
     raise CondaBuildUserError("TESTS FAILED: " + os.path.basename(pkg))
-
-
-@deprecated(
-    "24.7",
-    "24.9",
-    addendum="`patchelf` is an explicit conda-build dependency on Linux so it will always be installed.",
-)
-def check_external():
-    if on_linux:
-        patchelf = external.find_executable("patchelf")
-        if patchelf is None:
-            raise CondaBuildUserError(
-                f"Did not find 'patchelf' in: {os.pathsep.join(external.dir_paths)} "
-                f"'patchelf' is necessary for building conda packages on Linux with "
-                f"relocatable ELF libraries.  You can install patchelf using conda install "
-                f"patchelf."
-            )
 
 
 def build_tree(