+++ tests/test_build.py
@@ -15,12 +15,12 @@ from pathlib import Path
 from typing import TYPE_CHECKING
 
 import pytest
-from conda.common.compat import on_linux, on_win
+from conda.common.compat import on_win
 
 from conda_build import api, build
 from conda_build.exceptions import CondaBuildUserError
 
-from .utils import get_noarch_python_meta, metadata_dir, metadata_path
+from .utils import get_noarch_python_meta, metadata_dir
 
 if TYPE_CHECKING:
     from conda_build.config import Config
@@ -338,21 +338,6 @@ def test_guess_interpreter(
         assert build.guess_interpreter(script) == interpreter
 
 
-def test_check_external():
-    with pytest.deprecated_call():
-        build.check_external()
-
-
-@pytest.mark.skipif(not on_linux, reason="pathelf is only available on Linux")
-def test_check_external_user_error(mocker: MockerFixture) -> None:
-    mocker.patch(
-        "conda_build.os_utils.external.find_executable",
-        return_value=None,
-    )
-    with pytest.raises(CondaBuildUserError):
-        build.check_external()
-
-
 @pytest.mark.parametrize("readme", ["README.md", "README.rst", "README"])
 def test_copy_readme(testing_metadata: MetaData, readme: str):
     testing_metadata.meta["about"]["readme"] = readme
@@ -362,14 +347,6 @@ def test_copy_readme(testing_metadata: MetaData, readme: str):
     Path(testing_metadata.config.work_dir, readme).touch()
     build.copy_readme(testing_metadata)
     assert Path(testing_metadata.config.info_dir, readme).exists()
-
-
-def test_construct_metadata_for_test_from_recipe(testing_config: Config) -> None:
-    with pytest.warns(FutureWarning):
-        build._construct_metadata_for_test_from_recipe(
-            str(metadata_path / "test_source_files"),
-            testing_config,
-        )
 
 
 @pytest.mark.skipif(not on_win, reason="WSL is only on Windows")